[ re #8 , gallais/idris-tparsec#6 ] Position & Annotation support
Towards support for tracking the position of the parser in the source
file as well as putting annotations on subcomputations.

The idea is that both of these things can be recovered when failing
thus providing more precise error message.

The new things are mainly in (with documentation!):
* Text.Parser.Types
* Text.Parsre.Instruments

At the moment I'm stuck on porting the examples because the standard
library's `StateT` is a mere definition, not a record wrapper. This
makes getting Agda to solve instance problem sof the form
`RawMonad (StateT S M)` virtually impossible.

Looks like I'm going to have to roll out my version of `Category.Monad`
where I'll be using records rather than type synonyms...